+++ CHANGELOG.md
@@ -10,6 +10,7 @@ backwards compatibility.
 * Added sinc waveform into the default bank.
 * Add FM wavetable bank. Only to pure data for now.
 * Add electric guitar wavetable bank. Only to pure data for now.
+* Add distorted wavetable bank. Only to pure data for now.
 
 ## 0.5.0